SDMO: Added skip for already mined repos

+++ github_analyze_refactor_miner_reports.py
@@ -10,7 +10,7 @@
 import shutil
 from git import Repo  # pip install gitpython
 from pydriller import Repository #pip install pydriller
-from common import relative_to_absolute, read_csv, write_csv, read_json, makedirs_helper, get_repo_name, write_json
+from common import relative_to_absolute, read_csv, write_csv, read_json, makedirs_helper, get_repo_name, write_json, file_exists
 
 #Extract all the commits that have been deemed to have refactorings
 def get_refactored_commits(mined_repo):
@@ -56,28 +56,31 @@ mined_repos = read_csv(relative_to_absolute("mining_index.csv"), ",", {"source_g
 #process repos
 for mr in mined_repos:
     print("Mining details from: " + mr["source_git"])
-    #Get all refactored commits
-    refactored_commits = get_refactored_commits(mr)
-    #Mine wanted data from the repo
-    mined_commits = mine_details_from_repo(mr, refactored_commits)
-    print("Outputting submission for: " + mr["source_git"])
-
-    #output to submission folder
-    #Create folder for repo
+    #check if we want to skip this
     repo_submission_folder = submission_folder + "/" + get_repo_name(mr["source_git"])
-    makedirs_helper(repo_submission_folder)
-    #Copy rminer output
-    shutil.copyfile(mr["mining_report"], repo_submission_folder + "/rminer-output.json")
-    #write refactor commit message file
-    write_json(
-        repo_submission_folder + "/rcommit-messages.json", 
-        [{"commit_hash" : commit["hash"], "commit_message": commit["msg"]} for commit in mined_commits]
-    )
-    #write diff json
-    write_json(
-        repo_submission_folder + "/rcommit-diffs.json",
-        [{"commit_hash": commit["hash"], "previous_commit_hash": commit["previous_hash"], "diff": commit["diff"]} for commit in mined_commits]
-    )
+    if not file_exists(repo_submission_folder):
+        #Get all refactored commits
+        refactored_commits = get_refactored_commits(mr)
+        #Mine wanted data from the repo
+        mined_commits = mine_details_from_repo(mr, refactored_commits)
+        print("Outputting submission for: " + mr["source_git"])
+
+        #output to submission folder
+        #Create folder for repo
+
+        makedirs_helper(repo_submission_folder)
+        #Copy rminer output
+        shutil.copyfile(mr["mining_report"], repo_submission_folder + "/rminer-output.json")
+        #write refactor commit message file
+        write_json(
+            repo_submission_folder + "/rcommit-messages.json", 
+            [{"commit_hash" : commit["hash"], "commit_message": commit["msg"]} for commit in mined_commits]
+        )
+        #write diff json
+        write_json(
+            repo_submission_folder + "/rcommit-diffs.json",
+            [{"commit_hash": commit["hash"], "previous_commit_hash": commit["previous_hash"], "diff": commit["diff"]} for commit in mined_commits]
+        )
 
     #Create index for easy access in part 2
     mr["commit_messages"] = repo_submission_folder + "/rcommit-messages.csv"
